#VMS-2248 added std::thread replacement (nx::utils::thread). Needed to run on nx1
#VMS-2248 avoiding instanciating std::random_device often since it may throw
#VMS-2248 fixed exception specification
#VMS-2248 removed duplicate randomBuffer implementation. single std::random_device instance is used
#VMS-2248 removed redundant inheritance
#VMS-2248 review fixes
#VMS-2447 fixed unsynchronized variable access
CLOUD-237 review fixes
CLOUD-324 review fixes
Fix compiling after replacing qFill to std::fill.
Fix linux compile error
VMS-2342 QnTransactionTcpProcessor temporary socket mode fix
VMS-2391 nx::stun::AsyncClient reconnect tuning
VMS-2420 fix possible race in AxisResource
VMS-2420 replace deprecated Qt Algorithms
VMS-326 SSL refactoring, tests and review fixes
VMS-326 SSL sockets tests and fixes
VMS-326 cloud_connect_test_util --ssl support (tested) and mediaserver cloud ssl
VMS-326 review fix
avoiding std::async since it is buggy in GNU STL
avoiding std::thread since it is buggy in GNU STL
fixed compile error on gcc 4.8
fixed compile errors and warningsa
fixed compile warnings/errors
fixed nx1 build
fixing unit tests on arm
future.h moved to common_libs/nx_utils/src/nx/utils/std/
getting rid of std::future/promise in unit tests
no std::async anymore
static_unique_ptr_cast moved to <nx/utils/type_utils.h>
tomporarry cloud_db_ut comment for -Dut

<==== Mercurial data ====>
Branch: dev_3.0.0_gui

+++ unit_tests/nx_network_ut/src/pcp/router_ut.cpp
@@ -28,7 +28,7 @@ TEST(PcpRouter, StaticMap)
 
     EXPECT_TRUE(TestRouter::parseMapResponse(QByteArray::fromHex(response), mapping));
     EXPECT_EQ(mapping.external.toString().toUtf8(), QByteArray("32.43.12.43:4660"));
-    EXPECT_GT(mapping.lifeTime, 0);
+    EXPECT_GT(mapping.lifeTime, (size_t)0);
 }
 
 TEST(PcpRouter, DISABLED_RealMap)
